pyproject.toml
[project]
name = "stock-portfolio-monitoring"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dash[diskcache] (>=3.2.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "dash-ag-grid (>=31.3.1,<32.0.0)",
    "dash-mantine-components (>=1.2.0,<2.0.0)",
    "dash-iconify (>=0.1.2,<0.2.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pymongo (>=4.12.1,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "diskcache (>=5.6.3,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"
ruff = "^0.12.9"

